+++ pkg/agent/manager/cache/bundle_cache.go
@@ -2,8 +2,11 @@ package cache
 
 import (
 	"github.com/imkira/go-observer"
+	"github.com/spiffe/go-spiffe/v2/bundle/spiffebundle"
 	"github.com/spiffe/go-spiffe/v2/spiffeid"
 )
+
+type Bundle = spiffebundle.Bundle
 
 type BundleCache struct {
 	trustDomain spiffeid.TrustDomain
