+++ pkg/agent/manager/cache/lru_cache_test.go
@@ -19,6 +19,18 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+var (
+	trustDomain1       = spiffeid.RequireTrustDomainFromString("domain.test")
+	trustDomain2       = spiffeid.RequireTrustDomainFromString("otherdomain.test")
+	bundleV1           = spiffebundle.FromX509Authorities(trustDomain1, []*x509.Certificate{{Raw: []byte{1}}})
+	bundleV2           = spiffebundle.FromX509Authorities(trustDomain1, []*x509.Certificate{{Raw: []byte{2}}})
+	bundleV3           = spiffebundle.FromX509Authorities(trustDomain1, []*x509.Certificate{{Raw: []byte{3}}})
+	otherBundleV1      = spiffebundle.FromX509Authorities(trustDomain2, []*x509.Certificate{{Raw: []byte{4}}})
+	otherBundleV2      = spiffebundle.FromX509Authorities(trustDomain2, []*x509.Certificate{{Raw: []byte{5}}})
+	defaultX509SVIDTTL = int32(700)
+	defaultJwtSVIDTTL  = int32(800)
+)
+
 func TestLRUCacheFetchWorkloadUpdate(t *testing.T) {
 	cache := newTestLRUCache(t)
 	// populate the cache with FOO and BAR without SVIDS
@@ -1116,4 +1128,105 @@ func subscribeToWorkloadUpdates(t *testing.T, cache *LRUCache, selectors []*comm
 	subscriber, err := cache.subscribeToWorkloadUpdates(context.Background(), selectors, nil)
 	assert.NoError(t, err)
 	return subscriber
+}
+
+func distinctSelectors(id, n int) []*common.Selector {
+	out := make([]*common.Selector, 0, n)
+	for i := 0; i < n; i++ {
+		out = append(out, &common.Selector{
+			Type:  "test",
+			Value: fmt.Sprintf("id:%d:n:%d", id, i),
+		})
+	}
+	return out
+}
+
+func assertNoWorkloadUpdate(t *testing.T, sub Subscriber) {
+	select {
+	case update := <-sub.Updates():
+		assert.FailNow(t, "unexpected workload update", update)
+	default:
+	}
+}
+
+func assertAnyWorkloadUpdate(t *testing.T, sub Subscriber) {
+	select {
+	case <-sub.Updates():
+	case <-time.After(time.Minute):
+		assert.FailNow(t, "timed out waiting for any workload update")
+	}
+}
+
+func assertWorkloadUpdateEqual(t *testing.T, sub Subscriber, expected *WorkloadUpdate) {
+	select {
+	case actual := <-sub.Updates():
+		assert.NotNil(t, actual.Bundle, "bundle is not set")
+		assert.True(t, actual.Bundle.Equal(expected.Bundle), "bundles don't match")
+		assert.Equal(t, expected.Identities, actual.Identities, "identities don't match")
+	case <-time.After(time.Minute):
+		assert.FailNow(t, "timed out waiting for workload update")
+	}
+}
+
+func makeBundles(bundles ...*Bundle) map[spiffeid.TrustDomain]*Bundle {
+	out := make(map[spiffeid.TrustDomain]*Bundle)
+	for _, bundle := range bundles {
+		td := spiffeid.RequireTrustDomainFromString(bundle.TrustDomain().IDString())
+		out[td] = bundle
+	}
+	return out
+}
+
+func makeX509SVIDs(entries ...*common.RegistrationEntry) map[string]*X509SVID {
+	out := make(map[string]*X509SVID)
+	for _, entry := range entries {
+		out[entry.EntryId] = &X509SVID{}
+	}
+	return out
+}
+
+func makeRegistrationEntry(id string, selectors ...string) *common.RegistrationEntry {
+	return &common.RegistrationEntry{
+		EntryId:     id,
+		SpiffeId:    "spiffe://domain.test/" + id,
+		Selectors:   makeSelectors(selectors...),
+		DnsNames:    []string{fmt.Sprintf("name-%s", id)},
+		X509SvidTtl: defaultX509SVIDTTL,
+		JwtSvidTtl:  defaultJwtSVIDTTL,
+	}
+}
+
+func makeRegistrationEntryWithTTL(id string, x509SVIDTTL int32, jwtSVIDTTL int32, selectors ...string) *common.RegistrationEntry {
+	return &common.RegistrationEntry{
+		EntryId:     id,
+		SpiffeId:    "spiffe://domain.test/" + id,
+		Selectors:   makeSelectors(selectors...),
+		DnsNames:    []string{fmt.Sprintf("name-%s", id)},
+		X509SvidTtl: x509SVIDTTL,
+		JwtSvidTtl:  jwtSVIDTTL,
+	}
+}
+
+func makeRegistrationEntries(entries ...*common.RegistrationEntry) map[string]*common.RegistrationEntry {
+	out := make(map[string]*common.RegistrationEntry)
+	for _, entry := range entries {
+		out[entry.EntryId] = entry
+	}
+	return out
+}
+
+func makeSelectors(values ...string) []*common.Selector {
+	var out []*common.Selector
+	for _, value := range values {
+		out = append(out, &common.Selector{Type: "test", Value: value})
+	}
+	return out
+}
+
+func makeFederatesWith(bundles ...*Bundle) []string {
+	var out []string
+	for _, bundle := range bundles {
+		out = append(out, bundle.TrustDomain().IDString())
+	}
+	return out
 }
