+++ pkg/agent/manager/cache/util.go
@@ -11,3 +11,9 @@ func sortEntriesByID(entries []*common.RegistrationEntry) {
 		return entries[a].EntryId < entries[b].EntryId
 	})
 }
+
+func sortIdentities(identities []Identity) {
+	sort.Slice(identities, func(a, b int) bool {
+		return identities[a].Entry.EntryId < identities[b].Entry.EntryId
+	})
+}
