+++ pkg/agent/manager/cache/workload.go
@@ -1,0 +1,43 @@
+package cache
+
+import (
+	"crypto"
+	"crypto/x509"
+
+	"github.com/spiffe/go-spiffe/v2/bundle/spiffebundle"
+	"github.com/spiffe/go-spiffe/v2/spiffeid"
+	"github.com/spiffe/spire/proto/spire/common"
+)
+
+type Selectors []*common.Selector
+
+// Identity holds the data for a single workload identity
+type Identity struct {
+	Entry      *common.RegistrationEntry
+	SVID       []*x509.Certificate
+	PrivateKey crypto.Signer
+}
+
+// Update holds information for an SVIDs update to the cache.
+type UpdateSVIDs struct {
+	// X509SVIDs is a set of updated X509-SVIDs that should be merged into
+	// the cache, keyed by registration entry id.
+	X509SVIDs map[string]*X509SVID
+}
+
+// WorkloadUpdate is used to convey workload information to cache subscribers
+type WorkloadUpdate struct {
+	Identities       []Identity
+	Bundle           *spiffebundle.Bundle
+	FederatedBundles map[spiffeid.TrustDomain]*spiffebundle.Bundle
+}
+
+func (u *WorkloadUpdate) HasIdentity() bool {
+	return len(u.Identities) > 0
+}
+
+// X509SVID holds onto the SVID certificate chain and private key.
+type X509SVID struct {
+	Chain      []*x509.Certificate
+	PrivateKey crypto.Signer
+}
